+++ shenyu-plugin/shenyu-plugin-httpclient/src/test/java/org/apache/shenyu/plugin/httpclient/HttpClientPropertiesTest.java
@@ -124,6 +124,7 @@ public class HttpClientPropertiesTest {
         HttpClientProperties httpClientProperties = new HttpClientProperties();
         HttpClientProperties.Pool pool = httpClientProperties.getPool();
         Assertions.assertEquals(HttpClientProperties.Pool.PoolType.FIXED, pool.getType());
+        Assertions.assertEquals(3000L, pool.getAcquireTimeout());
         pool.setAcquireTimeout(ConnectionProvider.DEFAULT_POOL_ACQUIRE_TIMEOUT);
         pool.setMaxConnections(ConnectionProvider.DEFAULT_POOL_MAX_CONNECTIONS);
         pool.setName("name");